SecondRankTensor => float instead of double

+++ Examples/AverageTensorImages.cxx
@@ -8,7 +8,7 @@ template <unsigned int ImageDimension>
 int AverageTensorImages(unsigned int argc, char *argv[])
 {
   // typedef itk::Vector<float,6> TensorType;
-  typedef itk::SymmetricSecondRankTensor<double, 3> TensorType;
+  typedef itk::SymmetricSecondRankTensor<float, 3> TensorType;
 
   typedef itk::Image<TensorType, ImageDimension>       ImageType;
   typedef itk::ImageRegionIteratorWithIndex<ImageType> IteratorType;
